EASY-794 Add Bootstrap to EASY.
Note: reset.css is included in Bootstrap so we no
longer need it in EASY WebUI. reset.css was still
referenced from dans-wicket RESTpage.html,
however that page doesn't really need it and it
was decided to remove reset.css altogether.
I've also decided not to add Bootstrap to
RESTpage.html as I think it adds unnecessary bulk
and these REST pages don't really need it to
display well (and it's supposed to be a machine
interface). Incidentally these REST pages might be
removed in the future anyway (as said by Jan).

Note2: pulled in Bootstrap & JQuery instead of
using CDN links as we are unsure of the safety and
reliability of using this method.

Change-Id: I3bfbffffc67d42c82958065de03a0b0f55a3b95a