EASY-306 Remove YUI, unused javascript, unused CSS.
NOTE: THIS COMMIT DEPENDS ON COMMIT
fb387140bbe5eda69fa493156e405f8b22ccc738 (EASY-798
Replace the original help-popup with a Bootstrap
modal). MERGE THE fb387 COMMIT BEFORE THIS COMMIT.

* Remove unused CSS: there was a CSS class that was never used and
used a YUI image;
* Remove unused Javascript: there were some unused versions of JQuery
present in the code-base, vtip.js which was never referenced and HelpPopup.js
which is no longer used;
* Remove YUI because with the replacement of the original help pop-up with a
Bootstrap modal it is no longer used anywhere in the code.

Change-Id: I9100872528f03b0ad9df93b7693addc2ce056828