EASY-730 Add Fontawesome.
Add Fontawesome for easy icon usage.
From the Fontawesome website:
"Font Awesome gives you scalable vector icons that can instantly be
customized — size, color, drop shadow, and anything that can be done
with the power of CSS."

Bootstrap has glyphicons, but this is a very limited set.
Because the DANS website also uses Fontawesome I have decided to
use it for EASY as well.

Change-Id: Idfb81ff46b169ae657afdbefeb8dea5be01b5bd6