NO-JIRA-ISSUE: fix TestLoginPage ForgottenPassword unit test.
The string used in the test to tell the user that
an e-mail has been sent was not the string that
was actually used. In fact that string was used
nowhere else in the code except the test.
I've removed the string from the properties file
and updated the unit test.

JP: made test less maintenance sensitive

Change-Id: Ica28b433a99d50a1590f269ea3ea4e1047198df2