NO JIRA ISSUE. ClassCastException under Java 6.
The line *was*

  final List<Object[]> rows = query.list();
  
Even though what query.list() returned run-time was *not* a list of
Object[] but of single values, this line did not provoke the error in
Java 6. Only when the line:

  final T value = (T) row 
  
was executed a ClassCastException occurred, as Java 6 seemed to remember
that row should somehow be an Object[] (even thought is wasn't). Java 7
seemed to correctly use the actual type of row and thus did not throw
the ClassCastException.

Unfortunately the Hibernate API is not very type-safe. We are using an
outdated version, but even the latest version does not use type
parameters here.

http://docs.jboss.org/hibernate/orm/4.3/javadocs/org/hibernate/Query.html#list%28%29


Change-Id: I0ed9ba95d6335237cac522f00fd9c790f4166c62