EASY-730 Replace 'width="20%"' on table cells in UserInfoEditPanel with Bootstrap classes.
I used the method of assigning Bootstrap column
classes to <col> elements to specify the column
widths of a table which seemed the most 'standard'
and practical method (see resources below).

Note: it turns out <col> and <colgroup> are valid
elements in HTML4.
Therefore the use of <col> to specify the column
widths of a table is preferred.
Also note that HTML5 no longer supports the 
'width' attribute on <col>.
We use a Bootstrap class instead so this is not a
problem.

Resources:
* Table column widths in Bootstrap tables: http://ericsaupe.com/custom-column-widths-in-bootstrap-tables/
* <col> in HTML4: http://www.w3.org/TR/html4/struct/tables.html#h-11.2.4
* <col in HTML5: http://www.w3.org/TR/html5/tabular-data.html#the-col-element

Change-Id: Ibf40e602fec346e1a322c2dc440b0c6a4c905cf0