EASY-730 Improve the help-popups: don't show trigger when no help text.
Only show the help-text trigger when there is actually content to
display. Previously a trigger button was always visible which meant the
user could only find out after clicking it that there was really no help
text which is a bad user experience.

I ran into a Wicket 1.4 bug with this.

I originally just put a
<wicket:enclosure child="popupHTML"></wicket:enclosure>
around the HTML (just inside the <wicket:panel>),
but this didn't work because of the bug.

Instead I had to do the enclosure on 'popupModal'
(<wicket:enclosure child="popupModal") and define
the 'isVisible' method of the 'popupModal'
WebMarkupContainer in the Java code.

This bug is fixed in Wicket 1.5:
https://issues.apache.org/jira/browse/WICKET-2636

Change-Id: I3ff22b091a1a35b36936cb5a4f07ef3c591c1387