EASY-730 Put managementbar as drop down in main navigation.
There was an issue where an InlineEnclosure (child="managementBarPanel")
on a <li> wouldn't work and a normal Enclosure around the <li> would.
This was fixed by manually setting visibility of the managementBarPanel
depending on the session user. This is not a very neat solution as the
managementBarPanel is already hidden by the CodedAuthz, but necessary to
make the InlineEnclosure work and it was already used for other
navigation items as well.

Change-Id: I182fcca29b49c7ba70662b7e17aaff25bacee57e