EASY-730 Improve the deposit into page with more information.
* Add a 3-step instruction about depositing data that refers to more
information on the DANS website;
* Fix tests;
* Remove the necessity to be logged in for the deposit intro page so we
can give new users all the information they need before they sign up
(this information used to be on the homepage, although they now have
access to even more information). This also improves the user experience
for users that have been referred from the DANS website (they are no
longer confronted with a login/register screen when they click on
'deposit' on the DANS website).

Change-Id: I2cb2140baf60e01d4d87fcfa938ebae52313d0e0