EASY-730 Remove unimplemented facet search dropdown from search bar.
Nice-to-have feature that we probably don't have time for before the
upcoming release.
Would be great to have this implemented in the future though. It would
make EASY look more like Narcis and it would be a pretty cool feature.

The idea is that you can perform your search query on a preselected
facet (the audiences facet would make the most sense) (see Narcis search
bar for an example).

Change-Id: Ifc3e47902f59bbb2efacf63aa0ed484f88b47aa7