EASY-909 Breid task-add-doi-to-old-datasets uit toegevoegen/updaten van in Fedora bestaande DOI's
A new method added that will send datasets to datacite to be either
created or updated.  

It is possible that the already created DOIs in Fedora have not been
successfully created in Datacite, and therefor we now send also the
datasets with a DOI.  When datacite receives a dataset with a DOI that
already exists in datacite registry, it updates the entry instead of
creating a new one. 

Change-Id: I28dcf09b3d34c7c4dc6fa88abec4776b605677c9