Made the namespaces to set for the urn and doi in easy-pid-generator
configurable.

In production during the month of September the DOI's assigned were
DOI's in the test namespace (10.5072). This mistake was caused by the
fact that the namespaces in the easy-pid-generator config file could not
be configured by easy-dtap. Fixed this and set the namespace in the
production vars file for easy-dtap to 10.17026/dans- (the correct DANS
namespace and prefix).

Change-Id: Ide850e0950916d01af9674c5d040bfd819dfe74b