TA-13 Omzetten Access Rights van datasets en de accessibility van de bestanden
- For the datasets with the identifiers given in a file, this task
changes dataset status from open-access-for-registered-users to
open-access.  Access rights of the files belonging to these datasets are
changed from known to anonymous. If the original dataset status is
something else than open-access-for-registered-users, it will not be
changed.  The same applies for the files; file access rights are changed
only for the files where the original rights is 'known'.
- Two output files are written; one including the pid's of the changed
datasets, one including information about the changed dataset files.

First version

Change-Id: Iba132ca514bb9c718db0854375a90c9431982695