EASY-988 Maximaal DansDatasetMetadata.xml voorbeeld
easy-ddm:

- new test to validate public example of easy-schema
- old test: don't catch exceptions in unit tests unless expected

easy-sword:

Restored test that converts src/test/resources/demoDDM.xml
into target/demoDDM/emd.* in xml and html format,
the html format appears in the license document of a deposit.
This should (but might not) be the same content as the second tab of the web-ui.

The OfflineValidator still needed online XSDs indirectly
so used an assume to check availability.

Change-Id: I1134522697699b8fa43f6aa84bd88ccbd98427d2